+++ Kayttoliittyma/src/main/java/com/ot1/kayttoliittyma/UusiLaskuController.java
@@ -1,5 +1,7 @@
 package com.ot1.kayttoliittyma;
 
+import javafx.collections.FXCollections;
+import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
@@ -18,13 +20,33 @@ public class UusiLaskuController {
     @FXML
     private DatePicker laskuErapaivaDatePicker;
     @FXML
-    private ChoiceBox asiakasChoiceBox;
+    private ChoiceBox <String> asiakasChoiceBox;
     @FXML
-    private ChoiceBox mokkiChoiceBox;
+    private ChoiceBox <String >mokkiChoiceBox;
     @FXML
-    private ChoiceBox paivatChoiceBox;
+    private ChoiceBox <String> paivatChoiceBox;
     @FXML
-    private ChoiceBox loppusiivousChoiceBox;
+    private ChoiceBox <String> loppusiivousChoiceBox;
+
+    @FXML
+    private void initialize() {
+        ObservableList<String> asiakasLista = FXCollections.observableArrayList("Asiakas 1", "Asiakas 2", "Asiakas 3");
+        ObservableList<String> mokkiLista = FXCollections.observableArrayList("Mökki1", "Mökki 2", "Mökki 3");
+        ObservableList<String> paivatLista = FXCollections.observableArrayList("1", "2", "3");
+        ObservableList<String> loppusiivousLista = FXCollections.observableArrayList("0", "1", "2");
+
+        asiakasChoiceBox.setValue("Asiakas 1");
+        asiakasChoiceBox.setItems(asiakasLista);
+
+        mokkiChoiceBox.setValue("Mökki 1");
+        mokkiChoiceBox.setItems(mokkiLista);
+
+        paivatChoiceBox.setValue("1");
+        paivatChoiceBox.setItems(paivatLista);
+
+        loppusiivousChoiceBox.setValue("0");
+        paivatChoiceBox.setItems(loppusiivousLista);
+    }
 
     //Toiminto koti-buttonille
     public void kotiButton(javafx.event.ActionEvent actionEvent) {
@@ -74,11 +96,35 @@ public class UusiLaskuController {
         System.out.println("Aloituspäivä: " + aloitusPaiva);
     }
 
-    //Eräpäicä datepicker toimimaan
+    //Eräpäivä datepicker toimimaan
     @FXML
     private void handleEraPaiva() {
         LocalDate eraPaiva = laskuErapaivaDatePicker.getValue();
         System.out.println("Eräpäivä: " + eraPaiva);
+    }
+    //Asiakas choiceBox toimimaan
+    @FXML
+    private void handleAsiakasCB() {
+        String valittuAsiakas = asiakasChoiceBox.getValue();
+        System.out.println("Valittu asiakas: " + valittuAsiakas);
+    }
+    //Mökki choiceBox toimimaan
+    @FXML
+    private void handleMokkiCB() {
+        String valittuMokki = mokkiChoiceBox.getValue();
+        System.out.println("Valittu mökki: " + valittuMokki);
+    }
+    //Majoitettujen päivien määrä toimimaan
+    @FXML
+    private void handlePaivatCB() {
+        String valitutPaivat = paivatChoiceBox.getValue();
+        System.out.println("Päivien määrä: " + valitutPaivat);
+    }
+    //Loppusiivous choiceBox toimiaan
+    @FXML
+    private void handleLoppusiivousCB() {
+        String loppusiivous = loppusiivousChoiceBox.getValue();
+        System.out.println("Loppusiivous: " + loppusiivous);
     }
 
 }
