+++ Kayttoliittyma/src/main/resources/ui/UusiLasku.fxml
@@ -6,7 +6,7 @@
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.text.Text?>
 
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="567.0" prefWidth="848.0" xmlns="http://javafx.com/javafx/23.0.1" xmlns:fx="http://javafx.com/fxml/1">
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="567.0" prefWidth="848.0" xmlns="http://javafx.com/javafx/23.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.ot1.kayttoliittyma.UusiLaskuController">
    <children>
       <Button fx:id="takaisinButton" layoutX="24.0" layoutY="24.0" mnemonicParsing="false" text="Takaisin" />
       <Button fx:id="kotiButton" layoutX="89.0" layoutY="24.0" mnemonicParsing="false" text="Koti" />
@@ -14,14 +14,14 @@
       <Text layoutX="128.0" layoutY="179.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Eräpäivä:" />
       <Text layoutX="129.0" layoutY="234.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Asiakas:" />
       <Text layoutX="127.0" layoutY="291.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Mökki:" />
-      <DatePicker fx:id="laskuPaivamaaraDatePicker" layoutX="298.0" layoutY="106.0" />
-      <DatePicker fx:id="laskuErapaivaDatePicker" layoutX="298.0" layoutY="162.0" />
-      <ChoiceBox fx:id="asiakasChoiceBox" layoutX="298.0" layoutY="217.0" prefHeight="25.0" prefWidth="174.0" />
-      <ChoiceBox fx:id="mokkiChoiceBox" layoutX="298.0" layoutY="274.0" prefHeight="25.0" prefWidth="174.0" />
-      <ChoiceBox fx:id="paivatChoiceBox" layoutX="298.0" layoutY="323.0" prefHeight="25.0" prefWidth="174.0" />
-      <Text layoutX="126.0" layoutY="340.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Majoitettujen päivien määrä:" />
+      <DatePicker fx:id="laskuPaivamaaraDatePicker" layoutX="298.0" layoutY="106.0" onAction="#handleAloitusPaiva"/>
+      <DatePicker fx:id="laskuErapaivaDatePicker" layoutX="298.0" layoutY="162.0" onAction="#handleEraPaiva"/>
+      <ChoiceBox fx:id="asiakasChoiceBox" layoutX="298.0" layoutY="217.0" prefHeight="25.0" prefWidth="174.0" onAction="#handleAsiakasCB"/>
+      <ChoiceBox fx:id="mokkiChoiceBox" layoutX="298.0" layoutY="274.0" prefHeight="25.0" prefWidth="174.0" onAction="#handleMokkiCB"/>
+      <ChoiceBox fx:id="paivatChoiceBox" layoutX="298.0" layoutY="323.0" prefHeight="25.0" prefWidth="174.0" onAction="#handlePaivatCB"/>
+      <Text layoutX="126.0" layoutY="340.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Majoitettujen päivien määrä:"/>
       <Text layoutX="126.0" layoutY="394.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Loppusiivous:" />
-      <ChoiceBox fx:id="loppusiivousChoiceBox" layoutX="298.0" layoutY="377.0" prefHeight="25.0" prefWidth="174.0" />
+      <ChoiceBox fx:id="loppusiivousChoiceBox" layoutX="298.0" layoutY="377.0" prefHeight="25.0" prefWidth="174.0" onAction="#handleLoppusiivousCB"/>
       <Button fx:id="tallennaButton" layoutX="123.0" layoutY="457.0" mnemonicParsing="false" prefHeight="25.0" prefWidth="90.0" text="Tallenna" />
    </children>
 </AnchorPane>
